api-gateway/api-gateway/requirements.txt
billiard==3.6.4.0
Flask==2.0.1
Flask-HTTPAuth==4.4.0
flask-jwt-extended==4.3.1
flask-opentracing==1.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
prompt-toolkit==3.0.20
pytz==2021.1
opentracing==2.0.0
requests-opentracing==0.2.0
six==1.16.0
sfx-jaeger-client==3.13.1b0.dev1
signalfx-tracing==0.0.15
SQLAlchemy==1.4.23
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
waitress==1.4.3
